#173 move runtime controller into runtime module
