+++ docs/EXTEND_STATE.md
@@ -18,7 +18,7 @@ Validation/deserialization:
 - `src/world/level.ts`
 
 Runtime state commit:
-- `src/world/world.ts`, `src/runtimeController.ts`, and callback wiring in `src/runtime/createRuntimeApp.ts`
+- `src/world/world.ts`, `src/runtime/runtimeController.ts`, and callback wiring in `src/runtime/createRuntimeApp.ts`
 
 ## State Extension Workflow
 
@@ -116,9 +116,9 @@ Required follow-up updates:
 - type updates in `src/world/types.ts`
 - deterministic default initialization in `src/world/state.ts` and `src/world/level.ts`
 - deterministic command handling in `src/world/world.ts` for `selectInventorySlot`
-- runtime command-indexed event emission in `src/runtimeController.ts` using an item-use resolver boundary
+- runtime command-indexed event emission in `src/runtime/runtimeController.ts` using an item-use resolver boundary
 - immutable event commit wiring in `src/runtime/createRuntimeApp.ts`
-- regression tests in `src/world/world.test.ts`, `src/runtimeController.test.ts`, and `src/input/keyboard.test.ts`
+- regression tests in `src/world/world.test.ts`, `src/runtime/runtimeController.test.ts`, and `src/input/keyboard.test.ts`
 
 ## Checklist
